# neoProject/src/week5/src/main/java/com/example/week5/dto/Request.java
package com.example.week5.dto;

import lombok.*;
import lombok.experimental.FieldDefaults;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@FieldDefaults(level = AccessLevel.PRIVATE)

public class Request {
    String email;
    String password;
}